update per pr comment
