--- src/core/public/chrome/ui/header/collapsible_nav_header.tsx
@@ -1,31 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-import { i18n } from '@osd/i18n';
-import React from 'react';
-import { EuiIcon, EuiFlexGroup, EuiFlexItem, EuiText, EuiCollapsibleNavGroup } from '@elastic/eui';
-
-export function CollapsibleNavHeader() {
-  const defaultHeaderName = i18n.translate(
-    'core.ui.primaryNav.workspacePickerMenu.defaultHeaderName',
-    {
-      defaultMessage: 'OpenSearch Dashboards',
-    }
-  );
-
-  return (
-    <EuiCollapsibleNavGroup>
-      <EuiFlexGroup>
-        <EuiFlexItem>
-          <EuiIcon type="logoOpenSearch" size="l" />
-        </EuiFlexItem>
-        <EuiFlexItem>
-          <EuiText>
-            <strong> {defaultHeaderName} </strong>
-          </EuiText>
-        </EuiFlexItem>
-      </EuiFlexGroup>
-    </EuiCollapsibleNavGroup>
-  );
-}
