+++ src/core/public/index.ts
@@ -350,12 +350,7 @@ export {
 
 export { __osdBootstrap__ } from './osd_bootstrap';
 
-export {
-  WorkspacesStart,
-  WorkspacesSetup,
-  WorkspacesService,
-  WorkspaceObservables,
-} from './workspace';
+export { WorkspacesStart, WorkspacesSetup, WorkspacesService } from './workspace';
 
 export {
   WorkspacePermissionMode,
