+++ src/core/public/workspace/index.ts
@@ -2,9 +2,4 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-export {
-  WorkspacesStart,
-  WorkspacesService,
-  WorkspacesSetup,
-  WorkspaceObservables,
-} from './workspaces_service';
+export { WorkspacesStart, WorkspacesService, WorkspacesSetup } from './workspaces_service';
